Moved checking of module dependencies to own files
Moved the implementation of determining if the Path order and module data product dependencies are consistent to its own files. This allows better sharing as well as the opportunity for standalone tests.
