+++ FWCore/Framework/src/PathsAndConsumesOfModules.cc
@@ -2,6 +2,8 @@
 
 #include "FWCore/Framework/interface/Schedule.h"
 #include "FWCore/Framework/src/Worker.h"
+#include "throwIfImproperDependencies.h"
+
 #include "FWCore/Utilities/interface/EDMException.h"
 
 #include <algorithm>
@@ -91,5 +93,110 @@ namespace edm {
         << "PathsAndConsumesOfModules::moduleIndex: Unknown moduleID\n";
     }
     return iter->second;
+  }
+  
+  //====================================
+  // checkForCorrectness algorithm
+  //
+  // The code creates a 'dependency' graph between all
+  // modules. A module depends on another module if
+  // 1) it 'consumes' data produced by that module
+  // 2) it appears directly after the module within a Path
+  //
+  // If there is a cycle in the 'dependency' graph then
+  // the schedule may be unrunnable. The schedule is still
+  // runnable if all cycles have at least two edges which
+  // connect modules only by Path dependencies (i.e. not
+  // linked by a data dependency).
+  //
+  //  Example 1:
+  //  C consumes data from B
+  //  Path 1: A + B + C
+  //  Path 2: B + C + A
+  //
+  //  Cycle: A after C [p2], C consumes B, B after A [p1]
+  //  Since this cycle has 2 path only edges it is OK since
+  //  A and (B+C) are independent so their run order doesn't matter
+  //
+  //  Example 2:
+  //  B consumes A
+  //  C consumes B
+  //  Path: C + A
+  //
+  //  Cycle: A after C [p], C consumes B, B consumes A
+  //  Since this cycle has 1 path only edge it is unrunnable.
+  //
+  //  Example 3:
+  //  A consumes B
+  //  B consumes C
+  //  C consumes A
+  //  (no Path since unscheduled execution)
+  //
+  //  Cycle: A consumes B, B consumes C, C consumes A
+  //  Since this cycle has 0 path only edges it is unrunnable.
+  //====================================
+  
+
+  void checkForModuleDependencyCorrectness(edm::PathsAndConsumesOfModulesBase const& iPnC,
+                                           bool iPrintDependencies) {
+    //Need to lookup names to ids quickly
+    std::map<std::string,unsigned int> moduleNamesToIndex;
+    for(auto const& description: iPnC.allModules()) {
+      moduleNamesToIndex.insert(std::make_pair(description->moduleLabel(),
+                                               description->id()));
+    }
+    
+    using namespace edm::graph;
+    
+    //If a module to module dependency comes from a path, remember which path
+    EdgeToPathMap edgeToPathMap;
+    
+    //for testing, state that TriggerResults is at the end of all paths
+    const std::string kTriggerResults("TriggerResults");
+    
+    const auto kInvalidIndex = std::numeric_limits<unsigned int>::max();
+    auto trItr = moduleNamesToIndex.find(kTriggerResults);
+    const unsigned int kTriggerResultsIndex = trItr != moduleNamesToIndex.end()? trItr->second : kInvalidIndex;
+    //determine the path dependencies
+    std::vector<std::string> pathNames = iPnC.paths();
+    {
+      
+      for(unsigned int pathIndex = 0; pathIndex != pathNames.size(); ++pathIndex) {
+        std::set<unsigned int> alreadySeenIndex;
+        
+        auto const& moduleDescriptions = iPnC.modulesOnPath(pathIndex);
+        unsigned int lastModuleIndex = kInvalidIndex;
+        for(auto const& description: moduleDescriptions) {
+          auto found = alreadySeenIndex.insert(description->id());
+          if(found.second) {
+            //first time for this path
+            unsigned int const moduleIndex = description->id();
+            if(lastModuleIndex  != kInvalidIndex ) {
+              edgeToPathMap[std::make_pair(moduleIndex,lastModuleIndex)].push_back(pathIndex);
+            }
+            lastModuleIndex = moduleIndex;
+          }
+        }
+        //Stick TriggerResults at end
+        if( (lastModuleIndex  != kInvalidIndex) and (kTriggerResultsIndex != kInvalidIndex) ) {
+          edgeToPathMap[std::make_pair(kTriggerResultsIndex,lastModuleIndex)].push_back(pathIndex);
+        }
+      }
+    }
+    {
+      //determine the data dependencies
+      for(auto const& description: iPnC.allModules()) {
+        unsigned int const moduleIndex = description->id();
+        auto const& dependentModules = iPnC.modulesWhoseProductsAreConsumedBy(moduleIndex);
+        for(auto const& depDescription: dependentModules) {
+          edgeToPathMap[std::make_pair(moduleIndex, depDescription->id())].push_back(kInvalidIndex);
+          if(iPrintDependencies) {
+            edm::LogAbsolute("ModuleDependency") << "ModuleDependency '" << description->moduleLabel() <<
+            "' depends on data products from module '" << depDescription->moduleLabel()<<"'";
+          }
+        }
+      }
+    }
+    graph::throwIfImproperDependencies(edgeToPathMap,pathNames,moduleNamesToIndex);
   }
 }
